+++ SmartDeviceLink/SDLProtocolMessageDisassembler.m
@@ -27,7 +27,7 @@ NS_ASSUME_NONNULL_BEGIN
 
     // How many messages do we need to create to hold that many bytes?
     // Note: this does NOT count the special first message which acts as a descriptor.
-    NSUInteger numberOfMessagesRequired = ceil((float)incomingPayloadSize / (float)(mtu - headerSize));
+    NSUInteger numberOfMessagesRequired = (NSUInteger)ceil((float)incomingPayloadSize / (float)(mtu - headerSize));
 
     // And how many data bytes go in each message?
     NSUInteger numberOfDataBytesPerMessage = mtu - headerSize;
@@ -50,7 +50,7 @@ NS_ASSUME_NONNULL_BEGIN
 
 
     // Create the middle messages (the ones carrying the actual data).
-    for (int n = 0; n < numberOfMessagesRequired - 1; n++) {
+    for (NSUInteger n = 0; n < numberOfMessagesRequired - 1; n++) {
         // Frame # after 255 must cycle back to 1, not 0.
         // A 0 signals last frame.
         UInt8 frameNumber = (n % 255) + 1;
